Create general metadata-change event
